// samples/web-csr/dressca-frontend/consumer/src/system-common/generated/api-client/models/basket-item-api-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca
 * ECサイトDressca
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DisplayItemSummaryApiModel } from './display-item-summary-api-model';

export interface BasketItemApiModel {
    'displayItem'?: DisplayItemSummaryApiModel;
    'displayItemId': string;
    'quantity': number;
    'subTotal': number;
    'unitPrice': number;
}