// samples/web-csr/dressca-frontend/consumer/src/system-common/generated/api-client/models/get-basket-items-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca
 * ECサイトDressca
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AccountApiModel } from './account-api-model';
// May contain unused imports in some cases
// @ts-ignore
import type { BasketItemApiModel } from './basket-item-api-model';

export interface GetBasketItemsResponse {
    'account'?: AccountApiModel;
    'basketItems'?: Array<BasketItemApiModel>;
    'buyerId': string;
    'deletedItemIds'?: Array<string>;
}